Fix: reference to nonexistant `inputFiles`
I'd like to pass a glob pattern into `walkSync.entries` because this function is the source of a large bottleneck in my ember build pipeline.  It would be nice if sassOptions.includePaths was translated into this option. This is probably the first of a few changes necessary to support passing that glob pattern. In any event, this looks like a bug as this.inputFiles is not specified anywhere.